virtiofs: Add AVF API to share directory paths
This patch adds the API to share directory path between host
and guest.

The path to be shared is taken as an input from vm_config.json;

```
    "sharedPath": [
        {
            "sharedPath": "/data/data/com.google.android.virtualization.terminal/files"
        }
    ],
```

The UID-GID mapping between host and guest is constructed on the fly
based on the active user.

Currently, terminal app's internal storage is set as default; but this
can be changed as required.

This is now disabled by default - the feature will be enabled once
all the dependencies related to crosvm lands.

Bug: 372171883
Test: Share terminal app internal storage to guest VM. Read "debian.log"
files inside the guest. Verify UID-GID mappings are set correctly to the
files.

Change-Id: I9f2e7d9167a00c8383523bbd238a48b06afa4c2e
Signed-off-by: Akilesh Kailash <akailash@google.com>