microdroid_manager: wait for failure reason to transmit
Writing only 16 bytes at a time makes the race condition less likely to
happen, but, theoretically, it isn't a complete fix. Instead, use
`tcdrain` to block until the serial port driver knows all the data has
been trasmitted.

In the very unlikely case that the failure reason is larger than the
pipe capacity, this could cause a deadlock, so the pipe read is moved
into a new thread.

I was not able to reproduce the race condition either before or after
this change.

Bug: 220071963
Test: packages/modules/Virtualization/android/vm/vm_shell.sh start-microdroid
Change-Id: I3626e942f4a4ec86422c40befb28ee34f2682f22