pvmfw: Move aarch64 exceptions handling to new `arch` module
Due to incoming changes in libvmbase and pvmfw that will introduce new CPU architecture, there is necessary to separate code basing on platform. To achieve this new module/directory `arch` is introduced by this commit. This new module should be used as split point between architectures - new platform specific code should be only added behind `arch` module and exposed in platform independent way to the rest of the program.

Bug: 362733888
Test: m pvmfw

Change-Id: Ife352453563232d8313061a1ada844b01011a9bc