add option for custom vCPU count
Only supported by VirtualMachineRawConfig.

It has been observed that >1 vCPU hurts performance in some cases when
not also setting the CPU affinities and performing other tweaks, as done
by MATCH_HOST, so we avoided adding this before. However, there are
probably use cases where it is the best option and we've been explicitly
asked by partners to add it.

Bug: 381281047
Test: m
Change-Id: Ic22c747d064e70505398eb438a210ef42b5d75e1