Skip: isNewInstanceTest nPVM/without Secretkeeper
Non protected VM using legacy Secret mechanism do not reliably implement
the isNewInstance. This is because, in case of legacy mechanism, pVM
would rely on pvmfw to set the appropriate DT property. But
non-protected VM, pvmfw is never run.

Note that is possible to implement this for nPVM - for e.g., by checking
if there is already existing saved data in Microdroid partition in
instance.img. We disable the test first as mitigation.

Test: NA
Bug: 389030806
Change-Id: I5510c56de90fd001a78249504a599df0f9cb5aa2