USE_RELAXED_MICRODROID_ROLLBACK_PROTECTION: also add i18 apex
This is a hack for this release. It will be removed once we've added a
@SystemApi to load additional apexes to Microdroid pVMs.

This patch also adds a simple test to assert that com.android.i18n APEX
is correctly mounted. However, this test is not enough as we should also
assert that payload can actually use libraries provided by the APEX.
This will be done in a follow up patch.

Bug: 378681279
Bug: 390557313
Test: atest MicrodroidTests
Test: presubmit
Change-Id: I6f7013f9e5c693e72d7a9151d78d06401e4159f6