pvmfw: Add fixed-digest RBP for desktop TEE VM
Define a reserved com.android.virt.name for the TEE VM and detect it in
pvmfw to perform fixed-digest rollback protection on platforms that ship
the VM or reject payloads with the name on platforms that don't.

Distinguish between the 2 by only enabling "platform_has_desktop_trusty"
for pvmfw_trusty, for now.

Bug: 392628867
Bug: 402505709
Test: m libpvmfw libpvmfw_desktop
Test: mmm packages/modules/Virtualization/guest/pvmfw
Flag: EXEMPT b/406977494
Change-Id: I5dc574c163041b09338d73b2f9d3fbbfee0114ef