test microdroid vendor images: set avb_hash_algorithm to "sha256"
When not specified the avb_hash_algorithm defaults to "sha1", which is
discouraged in Android. This also blocks the updated to microdroid
kernel prebuilts (the sha1 support was recently removed from the
microdroid_defconfig).

The aosp/50fb49cabe70cfbde6743acb18543cd14befb7a5 changes build system
to default avb_sha_algorithm to "sha256", however it might take some
time for that patch to be reviewed, hence manually setting the value
here.

Test: atest MicrodroidTests
Change-Id: I67fb4512b65dcc76843d0a426517b8ef9cc67339