pvmfw: fdt: Support u64 kernel & initrd ranges
As CrosVM might now generate 64-bit values for the ramdisk range [1],
and as pvmfw should not resize the property during its verification,
introduce DeviceTreeInteger to keep track of not only the value but also
the original size of an integer property received through the input DT
and use it for "linux,initrd-start" and "linux,initrd-end".

Use the same parsing code for "kernel-address" and "kernel-size", to
permit the same flexibility to the VMM when generating those values.

[1]: https://crrev.com/c/6241064

Test: m pvmfw
Change-Id: I8465f3bd9d4e175f72913e3e6a254ee626b74d7f