Make packages/modules/Virtualization tests available to the platform
Soong was accidentally allowing a corner case where modules that
were not supposed to be available to the platform (due to
apex_available being set to an apex and not including
"//apex_available:platform") were being made available to the
platform as long as no apex modules depended on them.  Move the
apex_available: ["com.android.virt"] from the defaults modules
to the rust_library modules so that it doesn't apply to the
rust_test modules.

Bug: 394664568
Test: compare out/soong/installs-${TARGET_PRODUCT}.mk
Change-Id: Idff0deb6063ad0bc4ce4a95c43f6fb4d19e57260