virtmgr: also support non-vendor tee services
We are going to reuse the teeServices field of the
VirtualMachineRawConfig to also handle cases of generic "system" tee
services. First such case (guest-ffa-proxy) will be added in the
follow-up patch.

This patch simply splits teeServices into "vendor" (name has vendor.
prefix) and "system" (the rest). The handling of the "vendor" tee
services will be partially delegated to the IVmCapabilitiesService HAL
in one of the follow-up patches, while handling of the "system" tee
services will be entirely done between virtmgr & crosvm.

Bug: 391774181
Bug: 360102915
Test: presubmit
Change-Id: I9ef85475204ed53a56dbf2c25661a7c0b0da3cb0