Increase the priority of microdroid_launcher process
When capturing traces inside Microdroid VM (e.g. by using perfetto), the tracing daemons and adbd / sh processes sometimes can prevent the payload from running, which skews the results of the tracing.

This patch attempts to minimize the effect of tracing by increasing the priority of the microdroid_launcher process. This is achieved by setting the nice value to -20 (lowest possible).

Test: boot microdroid & check nice value of microdroid_launcher
Change-Id: I0d1c41a803d41f11c1c58f5b85d3b8aa9eb8bca8