Start a VM in suspended state if vendor_tee_services were requested
If a VM requests access to vendor_tee_services, virtmngr will tell
crosvm to start a VM in a suspended state. Then it will get the vm_fd
that represents this VM, and pass it to the new HAL to grant access for
this VM to the tee services. After that VM will be resumed.

This patch just implements the "start VM suspended" and "resume VM"
parts. The get vm_fd and communication with the HAL will be implemented
in the follow up patches.

Bug: 360102915
Test: vm run-microdroid --tee-services vendor.test_pkvm_tee_service
Change-Id: I57df7547dd0245222af4d768396d4ae5fa9e15f2