pvmfw: Support QTVMs through missing DICE handover
As pvmfw-in-QTVM receives a configuration data that is dynamically
generated per VM by Gunyah's Resource Manager (RM) and as QTVM guests do
not rely on DICE, a missing DICE handover in the config data gets
interpreted as "QTVM mode", where AVB is skipped (because QTVMs rely on
TZ-enabled PIL verification) and RBP and DICE must be skipped as a
result, as they lack useful inputs.

This change should not affect other distributions of pvmfw, as the DICE
handover is mandatory and well-behaved pvmfw loaders (ABL) MUST provide
one, as per the requirements of Android 25Q2.

Test: m pvmfw
Bug: 393977894
Change-Id: I74c8b81b396ccb20f987ef152af72c73a213e3d6