virtmgr: adjust default swiotlb algorithm
This commit should be a no-op. The new algorithm gives the same answer
(14 MB) for a default microdroid VM config and the same or close or
other configs.

The new algorithm factors out of some of the predictable parts of
Linux's swiotlb usage. These will be the target of optimizations in
follow up commit.

Some parts take into account the page size now, but it is hardcoded to
4k because I haven't tested it with 16k mode yet.

Flag: NONE noop
Test: launch microdroid
Bug: 405127241
Change-Id: I2218177ccec98c7386dd78c6eb1ef1bf33f99e6f